\name{check}
\alias{check}
\title{
  Check Observations Against Formal Description
}
\description{
  Check if observations correspond to their formal descriptions
}
\usage{
check(object, data)
}
\arguments{
  \item{object}{an object of class \code{var} or \code{vars}}
  \item{data}{a \code{data.frame}}
}
\details{
  The function returns true of \code{data} matches the description
  in \code{object}.
}
