\name{variables-package}
\docType{package}
\alias{variables-package}
\alias{variables}
\title{General Information on the \pkg{variables} Package}
\description{
  The \pkg{variables} package offers a small collection of objects describing 
  conceptual variables and corresponding methods, for example for generating
  a grid of values for a (yet) unmeasured variable.

  The package was written to support the \pkg{basefun} and \pkg{mlt}
  packages and will be of limited use outside these packages.
}
\author{
  This package is authored by Torsten Hothorn <Torsten.Hothorn@R-project.org>.
}
\keyword{package}
